Added Footer component, Profile.js page, and re-organized components/pages folders for naming conventions
