+++ client/src/components/About/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-const About  = () => {
+const About = () => {
     return (
         <div>
             <h2>About us</h2>
