--- client/src/components/Home/index.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Home = () => {
-   <div></div>
-};
-
-export default Home;
